pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.10",
    "fastapi>=0.115.13",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
]
